requirements.txt
httpx==0.27.2
openai==1.53.0
python-dotenv==1.0.1
python-telegram-bot==21.6
python-dotenv==1.0.1
